+++ Changes.md
@@ -39,6 +39,7 @@ Improvements
 - 3Delight :
   - Added camera overscan support.
   - NSI scene description export format is now based on file extension - `.nsi` for binary and `.nsia` for ASCII.
+  - Added support for reading `dl:` and `user:` attributes from shaders.
 
 Fixes
 -----