+++ Changes.md
@@ -5,6 +5,7 @@ Improvements
 ------------
 
 - ArnoldShader : Added a colour space presets menu for the `image` shader.
+- CyclesShader : Added a colour space presets menu for the `image_texture` and `environment_texture` shaders (#5618).
 
 API
 ---