+++ Changes.md
@@ -41,6 +41,7 @@ Improvements
   - NSI scene description export format is now based on file extension - `.nsi` for binary and `.nsia` for ASCII.
   - Added support for reading `dl:` and `user:` attributes from shaders.
   - Added `importanceSampleFilter` plug to DelightOptions, providing denoiser-compatible output.
+  - Matched DelightOptions default values for `oversampling` and `shadingSamples` to 3Delight's own default values.
 
 Fixes
 -----
@@ -123,6 +124,7 @@ Breaking Changes
 - ImageStatsUI : Removed `postCreate()`.
 - 3Delight : Changed NSI scene description export with `.nsi` file extension from ASCII to binary (`.nsia` is used for ASCII now).
 - OSLShader : Output parameters are now loaded onto the `out` plug for all types (`surface`, `displacement` etc), not just `shader`.
+- DelightOptions : Changed default values for `oversampling` and `shadingSamples` plugs.
 
 Build
 -----